+++ CHANGELOG.md
@@ -29,6 +29,7 @@ To be merged with *[Unreleased] - 2020-11-24*
 ### Added
 * Build support for Python 3.9
 * Make target for building providers
+* Now possible to update the `terraform` property of Terrascript by adding it multiple times
 * Provider:
   - `esxi`
   - `helm`